demo_projects/arxiv/pyproject.toml
[project]
name = "arxiv-demo"
version = "0.1.0"
description = "Databricks AI demo for analyzing Arxiv papers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.3.1",
    "databricks-sdk>=0.73.0",
    "python-dotenv>=1.2.1",
    "databricks-sdk>=0.73.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
    "mlflow>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arxiv_demo"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
